+++ packages/babel-preset-expo/src/index.ts
@@ -1,13 +1,35 @@
-const lazyImportsBlacklist = require('./lazy-imports-blacklist');
+import { ConfigAPI, PluginItem, TransformOptions } from '@babel/core';
 
-let hasWarnedJsxRename = false;
+import { lazyImports } from './lazyImports';
 
-module.exports = function (api, options = {}) {
+type BabelPresetExpoPlatformOptions = {
+  useTransformReactJSXExperimental?: boolean;
+  disableImportExportTransform?: boolean;
+  // Defaults to undefined, set to something truthy to disable `@babel/plugin-transform-react-jsx-self` and `@babel/plugin-transform-react-jsx-source`.
+  withDevTools?: boolean;
+  // Defaults to undefined, set to `true` to disable `@babel/plugin-transform-flow-strip-types`
+  disableFlowStripTypesTransform?: boolean;
+  // Defaults to undefined, set to `false` to disable `@babel/plugin-transform-runtime`
+  enableBabelRuntime?: boolean;
+  // Defaults to `'default'`, can also use `'hermes-canary'`
+  unstable_transformProfile?: 'default' | 'hermes-canary';
+};
+
+export type BabelPresetExpoOptions = {
+  lazyImports?: boolean;
+  reanimated?: boolean;
+  jsxRuntime?: 'classic' | 'automatic';
+  jsxImportSource?: string;
+  web?: BabelPresetExpoPlatformOptions;
+  native?: BabelPresetExpoPlatformOptions;
+};
+
+function babelPresetExpo(api: ConfigAPI, options: BabelPresetExpoOptions = {}): TransformOptions {
   const { web = {}, native = {}, reanimated } = options;
 
   const bundler = api.caller(getBundler);
   const isWebpack = bundler === 'webpack';
-  let platform = api.caller(getPlatform);
+  let platform = api.caller((caller) => (caller as any)?.platform);
 
   // If the `platform` prop is not defined then this must be a custom config that isn't
   // defining a platform in the babel-loader. Currently this may happen with Next.js + Expo web.
@@ -15,29 +37,26 @@ module.exports = function (api, options = {}) {
     platform = 'web';
   }
 
-  const platformOptions =
+  const platformOptions: BabelPresetExpoPlatformOptions =
     platform === 'web'
       ? {
           // Only disable import/export transform when Webpack is used because
           // Metro does not support tree-shaking.
-          disableImportExportTransform: !!isWebpack,
+          disableImportExportTransform: isWebpack,
           ...web,
         }
       : { disableImportExportTransform: false, ...native };
 
   // Note that if `options.lazyImports` is not set (i.e., `null` or `undefined`),
   // `metro-react-native-babel-preset` will handle it.
-  const lazyImportsOption = options && options.lazyImports;
+  const lazyImportsOption = options?.lazyImports;
 
-  const extraPlugins = [];
-
-  if ('useTransformReactJsxExperimental' in platformOptions && !hasWarnedJsxRename) {
-    // https://github.com/expo/expo/pull/13945#pullrequestreview-724327024
-    hasWarnedJsxRename = true;
-    console.warn(
-      'Warning: useTransformReactJsxExperimental has been renamed to useTransformReactJSXExperimental (capitalized JSX) in react-native@0.64.0'
-    );
-  }
+  const extraPlugins: PluginItem[] = [
+    // `metro-react-native-babel-preset` configures this plugin with `{ loose: true }`, which breaks all
+    // getters and setters in spread objects. We need to add this plugin ourself without that option.
+    // @see https://github.com/expo/expo/pull/11960#issuecomment-887796455
+    [require.resolve('@babel/plugin-proposal-object-rest-spread'), { loose: false }],
+  ];
 
   // Set true to disable `@babel/plugin-transform-react-jsx`
   // we override this logic outside of the metro preset so we can add support for
@@ -59,6 +78,15 @@ module.exports = function (api, options = {}) {
     // Purposefully not adding the deprecated packages:
     // `@babel/plugin-transform-react-jsx-self` and `@babel/plugin-transform-react-jsx-source`
     // back to the preset.
+  }
+
+  const aliasPlugin = getAliasPlugin();
+  if (aliasPlugin) {
+    extraPlugins.push(aliasPlugin);
+  }
+
+  if (platform === 'web') {
+    extraPlugins.push(require.resolve('babel-plugin-react-native-web'));
   }
 
   return {
@@ -90,12 +118,11 @@ module.exports = function (api, options = {}) {
           disableImportExportTransform: platformOptions.disableImportExportTransform,
           lazyImportExportTransform:
             lazyImportsOption === true
-              ? (importModuleSpecifier) => {
+              ? (importModuleSpecifier: string) => {
                   // Do not lazy-initialize packages that are local imports (similar to `lazy: true`
                   // behavior) or are in the blacklist.
                   return !(
-                    importModuleSpecifier.includes('./') ||
-                    lazyImportsBlacklist.has(importModuleSpecifier)
+                    importModuleSpecifier.includes('./') || lazyImports.has(importModuleSpecifier)
                   );
                 }
               : // Pass the option directly to `metro-react-native-babel-preset`, which in turn
@@ -104,53 +131,38 @@ module.exports = function (api, options = {}) {
         },
       ],
     ],
-    plugins: [
-      getObjectRestSpreadPlugin(),
-      ...extraPlugins,
-      getAliasPlugin(),
-      [require.resolve('@babel/plugin-proposal-decorators'), { legacy: true }],
-      platform === 'web' && [require.resolve('babel-plugin-react-native-web')],
-      isWebpack && platform !== 'web' && [require.resolve('./plugins/disable-ambiguous-requires')],
-      require.resolve('@babel/plugin-proposal-export-namespace-from'),
 
+    plugins: [
+      ...extraPlugins,
+      // TODO: Remove
+      [require.resolve('@babel/plugin-proposal-decorators'), { legacy: true }],
+      require.resolve('@babel/plugin-proposal-export-namespace-from'),
       // Automatically add `react-native-reanimated/plugin` when the package is installed.
+      // TODO: Move to be a customTransformOption.
       hasModule('react-native-reanimated') &&
         reanimated !== false && [require.resolve('react-native-reanimated/plugin')],
-    ].filter(Boolean),
+    ].filter(Boolean) as PluginItem[],
   };
-};
+}
 
-function getAliasPlugin() {
-  const aliases = {};
-
-  if (hasModule('@expo/vector-icons')) {
-    aliases['react-native-vector-icons'] = '@expo/vector-icons';
+function getAliasPlugin(): PluginItem | null {
+  if (!hasModule('@expo/vector-icons')) {
+    return null;
   }
-
-  if (Object.keys(aliases).length) {
-    return [
-      require.resolve('babel-plugin-module-resolver'),
-      {
-        alias: aliases,
+  return [
+    require.resolve('babel-plugin-module-resolver'),
+    {
+      alias: {
+        'react-native-vector-icons': '@expo/vector-icons',
       },
-    ];
-  }
-  return null;
+    },
+  ];
 }
 
-/**
- * metro-react-native-babel-preset configures this plugin with `{ loose: true }`, which breaks all
- * getters and setters in spread objects. We need to add this plugin ourself without that option.
- * @see https://github.com/expo/expo/pull/11960#issuecomment-887796455
- */
-function getObjectRestSpreadPlugin() {
-  return [require.resolve('@babel/plugin-proposal-object-rest-spread'), { loose: false }];
-}
-
-function hasModule(name) {
+function hasModule(name: string): boolean {
   try {
     return !!require.resolve(name);
-  } catch (error) {
+  } catch (error: any) {
     if (error.code === 'MODULE_NOT_FOUND' && error.message.includes(name)) {
       return false;
     }
@@ -158,32 +170,22 @@ function hasModule(name) {
   }
 }
 
-function getPlatform(caller) {
-  return caller && caller.platform;
-}
-
-/**
- * Get the name of the `bundler`.
- *
- * @param {*} caller
- */
-function getBundler(caller) {
+/** Determine which bundler is being used. */
+function getBundler(caller: any) {
   if (!caller) return null;
-
-  const { bundler, name } = caller;
-
-  if (!bundler) {
-    if (name === 'metro') {
-      // This is a hack to determine if metro is being used.
-      return 'metro';
-    } else if (name === 'next-babel-turbo-loader') {
-      // NextJS 11
-      return 'webpack';
-    } else if (name === 'babel-loader') {
-      // expo/webpack-config, gatsby, storybook, and next.js <10
-      return 'webpack';
-    }
+  if (caller.bundler) return caller.bundler;
+  if (
+    // Known tools that use `webpack`-mode via `babel-loader`: `@expo/webpack-config`, Next.js <10
+    caller.name === 'babel-loader' ||
+    // NextJS 11 uses this custom caller name.
+    caller.name === 'next-babel-turbo-loader'
+  ) {
+    return 'webpack';
   }
-  // Perhaps we should add a check to log once when an unexpected bundler is being used.
-  return bundler || null;
+
+  // Assume anything else is Metro.
+  return 'metro';
 }
+
+export default babelPresetExpo;
+module.exports = babelPresetExpo;